[Minor] Some more speed improvements for base64 decoding

+++ src/libcryptobox/chacha20/chacha.c
@@ -27,7 +27,7 @@
 #include "chacha.h"
 #include "platform_config.h"
 
-extern unsigned long cpu_config;
+extern unsigned cpu_config;
 
 typedef struct chacha_impl_t {
 	unsigned long cpu_flags;
